added entity interaction to the API, also added mouse position to interaction methods

+++ src/main/java/de/ellpeck/rockbottom/api/entity/player/IInteractionManager.java
@@ -32,7 +32,7 @@ public interface IInteractionManager{
 
     int getPlaceCooldown();
 
-    int getMousedTileX();
+    double getMousedTileX();
 
-    int getMousedTileY();
+    double getMousedTileY();
 }
